move functions and add correct includes in main
