+++ agileProjectRepo/include/steeringFunctions.h
@@ -1,0 +1,18 @@
+
+#pragma once
+
+void stopEngines()
+{
+    for (auto &engine : engines)
+    {
+        engine.setVelocity(0);
+    }
+}
+
+void setEnginesVelocity(int velocity)
+{
+    for (auto &engine : engines)
+    {
+        engine.setVelocity(velocity);
+    }
+}
